# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Marketing Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all recurring task schedules
      parameters:
        - name: propertyId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new recurring task schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "201":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Bad request
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get a schedule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule (edit or pause/resume)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleBody"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "404":
          description: Not found
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reviews
...
components:
  schemas:
...
    Schedule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        agentId:
          type: integer
        agentName:
          type: string
        agentIcon:
          type: string
        agentColor:
          type: string
        propertyId:
          type: integer
        propertyName:
          type: string
        taskType:
          type: string
        inputPrompt:
          type: ["string", "null"]
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        dayOfWeek:
          type: ["integer", "null"]
          description: "0=Sunday, 1=Monday ... 6=Saturday (weekly only)"
        dayOfMonth:
          type: ["integer", "null"]
          description: "1-31 (monthly only)"
        hour:
          type: integer
          description: "Hour of day to run (0-23)"
        timezone:
          type: string
        status:
          type: string
          enum: [active, paused]
        nextRunAt:
          type: ["string", "null"]
        lastRunAt:
          type: ["string", "null"]
        lastTaskId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - agentId
        - agentName
        - agentIcon
        - agentColor
        - propertyId
        - propertyName
        - taskType
        - frequency
        - hour
        - timezone
        - status
        - createdAt
        - updatedAt
    CreateScheduleBody:
      type: object
      properties:
        name:
          type: string
        agentId:
          type: integer
        propertyId:
          type: integer
        taskType:
          type: string
        inputPrompt:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        dayOfWeek:
          type: integer
          description: "0=Sunday ... 6=Saturday (required when frequency=weekly)"
        dayOfMonth:
          type: integer
          description: "1-31 (required when frequency=monthly)"
        hour:
          type: integer
          description: "Hour of day (0-23, defaults to 9)"
        timezone:
          type: string
          description: "IANA timezone (defaults to America/New_York)"
      required:
        - name
        - agentId
        - propertyId
        - taskType
        - frequency
    UpdateScheduleBody:
      type: object
      properties:
        name:
          type: string
        taskType:
          type: string
        inputPrompt:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly]
        dayOfWeek:
          type: integer
        dayOfMonth:
          type: integer
        hour:
          type: integer
        timezone:
          type: string
        status:
          type: string
          enum: [active, paused]
...