# lib/db/src/schema/schedules.ts
import { pgTable, text, serial, timestamp, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";
import { agentsTable } from "./agents";
import { propertiesTable } from "./properties";
import { tasksTable } from "./tasks";

export const schedulesTable = pgTable("schedules", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  agentId: integer("agent_id").notNull().references(() => agentsTable.id),
  propertyId: integer("property_id").notNull().references(() => propertiesTable.id),
  taskType: text("task_type").notNull(),
  inputPrompt: text("input_prompt"),
  frequency: text("frequency").notNull(),
  dayOfWeek: integer("day_of_week"),
  dayOfMonth: integer("day_of_month"),
  hour: integer("hour").notNull().default(9),
  timezone: text("timezone").notNull().default("America/New_York"),
  status: text("status").notNull().default("active"),
  nextRunAt: timestamp("next_run_at", { withTimezone: true }),
  lastRunAt: timestamp("last_run_at", { withTimezone: true }),
  lastTaskId: integer("last_task_id").references(() => tasksTable.id),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
  updatedAt: timestamp("updated_at", { withTimezone: true }).notNull().defaultNow().$onUpdate(() => new Date()),
});

export const insertScheduleSchema = createInsertSchema(schedulesTable).omit({
  id: true,
  createdAt: true,
  updatedAt: true,
  nextRunAt: true,
  lastRunAt: true,
  lastTaskId: true,
});

export type InsertSchedule = z.infer<typeof insertScheduleSchema>;
export type Schedule = typeof schedulesTable.$inferSelect;